requirements.txt
flask==3.1.0
python-dotenv==1.0.1
gunicorn==23.0.0
line-bot-sdk==3.16.2
langchain==0.3.20
langchain-openai==0.3.8
langchain-google-genai==2.1.0